backend/agents/requirements.txt
fastapi
uvicorn
PyMuPDF
pytesseract
pdf2image
pdfminer.six
camelot-py[cv]
opencv-python
pandas
python-multipart
regex